pyproject.toml
[tool.poetry]
name = "capibara"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
django-phonenumber-field = "^8.0.0"
psycopg = "^3.1.19"
setuptools = "^70.1.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
